Refactor: stores
