+++ src/features/groups/type.ts
@@ -1,0 +1,26 @@
+export type Group = {
+  groupId: string;
+  groupName: string;
+  stocks: { byId: { [stockId: string]: string[] }; allIds: string[] };
+};
+
+export interface GroupsState {
+  groups: { byId: { [groupId: string]: Group }; allIds: string[] };
+  selectedGroupId: string;
+  nextGroupId: number;
+}
+
+export type AddGroupPayload = {
+  groupName: string;
+  selectedStocks: Group['stocks'];
+};
+
+export type DeletePurchaseItemFromGroupPayload = {
+  stockId: string;
+  purchasedId: string;
+};
+export type UpdateMainGroupPayload = {
+  type: 'stock' | 'purchase';
+  stockId: string;
+  purchasedId: string;
+};
